refactored long values

+++ src/repository/DonationRepository.java
@@ -168,7 +168,7 @@ public class DonationRepository {
     for (Parameter<?> parameter : query.getParameters()) {
       countQuery.setParameter(parameter.getName(), query.getParameterValue(parameter));
     }
-    return countQuery.getSingleResult().longValue();
+    return countQuery.getSingleResult();
   }  
 
   public List<Donation> getAllDonations() {
